+++ src/form/inlineRadioGroup_templ.go
@@ -39,7 +39,7 @@ func InlineRadioGroup(componentSettings InlineRadioGroupSettings) templ.Componen
 			templ_7745c5c3_Var1 = templ.NopComponent
 		}
 		ctx = templ.ClearChildren(ctx)
-		templ_7745c5c3_Err = templruntime.WriteString(templ_7745c5c3_Buffer, 1, "<!-- InlineRadioGroup --><div class=\"border-1 bg-neutral-50/7.5 mt-2 flex h-10 flex-col rounded border-neutral-50/5 px-2\">")
+		templ_7745c5c3_Err = templruntime.WriteString(templ_7745c5c3_Buffer, 1, "<!-- InlineRadioGroup --><div class=\"border-1 bg-neutral-50/7.5 flex h-10 flex-col rounded border-neutral-50/5 px-2\">")
 		if templ_7745c5c3_Err != nil {
 			return templ_7745c5c3_Err
 		}