+++ src/import/toolset/index.js
@@ -1,0 +1,10 @@
+window.UiToolset = {
+  CreateRandomPassword: createRandomPassword,
+  ResolveApiResponseDisplay: resolveApiResponseDisplay,
+  ToggleLoadingOverlay: toggleLoadingOverlay,
+  RegisterAlpineState: registerAlpineState,
+};
+
+registerAlpineState(() => {
+  window.UiToolset.JsonAjax = jsonAjax;
+});